Cleanup if not completely correctify on tap controller (#33)
