+++ Sources/App/configure.swift
@@ -20,22 +20,18 @@ public func configure(_ config: inout Config, _ env: inout Environment, _ servic
 
     services.register(SQLiteKarmaRepository.self)
     services.register(RoomController.self)
-    services.register(SQLiteOnTapRepository.self)
 
     // Configure a SQLite database
     let sqlite = try SQLiteDatabase(storage: .file(path: "karmaDB"))
-    let ontapDatabase = try SQLiteDatabase(storage: .file(path: "onTapDB"))
 
     // Register the configured SQLite database to the database config.
     var databases = DatabasesConfig()
     databases.add(database: sqlite, as: .sqlite)
-    databases.add(database: ontapDatabase, as: .onTap)
     services.register(databases)
 
     // Configure migrations
     var migrations = MigrationConfig()
     migrations.add(model: Karma.self, database: .sqlite)
-    migrations.add(model: Beer.self, database: .onTap)
     services.register(migrations)
 
     services.register(OnTapController.self)
@@ -44,11 +40,4 @@ public func configure(_ config: inout Config, _ env: inout Environment, _ servic
     services.register(APIKeyStorage.self)
     services.register(Slack.self)
     services.register(SlackListener.self)
-}
-
-extension DatabaseIdentifier where D: Database {
-
-    public static var onTap: DatabaseIdentifier<SQLiteDatabase> {
-        return DatabaseIdentifier<SQLiteDatabase>("com.allen.unique._ontap.database.id")
-    }
 }
